refactor: move PixiViewport to own dir
