+++ src/components/pixi/viewport/types.ts
@@ -1,0 +1,5 @@
+import { Viewport } from 'pixi-viewport';
+
+export type PixiViewportInstance = {
+  viewport: Viewport;
+};
